clean supabase config
