+++ src/app/brothers/active/page.tsx
@@ -2,7 +2,7 @@
 import { useState, useEffect } from 'react';
 import { akpsiColors } from '../../../styles/colors';
 import { akpsiFonts } from '../../../styles/fonts';
-import { createClientBrowser } from '../../../lib/supabase';
+import { createClient } from '../../../../supabase/client';
 import { usePathname } from 'next/navigation';
 
 
@@ -12,7 +12,7 @@ export default function ActiveBrothers() {
 
   useEffect(() => {
     const fetchImage = async () => {
-      const supabase = createClientBrowser();
+      const supabase = await createClient();
       const { data, error } = await supabase
         .from('actives-spring25')
         .select('image_path, name')
